Revert "alterações pontuais no front"
