+++ .gitignore
@@ -1,7 +1,6 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/.history
 /node_modules
 /.pnp
 .pnp.js
