+++ package.json
@@ -19,7 +19,7 @@
     "chart": "^0.1.2",
     "chart.js": "^2.9.3",
     "icomoon-react": "^1.0.1",
-    "lodash": "^4.17.19",
+    "lodash": "^4.17.15",
     "moment": "^2.27.0",
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
@@ -28,7 +28,7 @@
     "react-redux": "^7.2.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.1",
+    "react-scripts": "3.0.1",
     "redux": "^4.0.5",
     "toastr": "^2.1.4",
     "typescript": "3.5.3"
