+++ src/App/Data/SidebarMenu.ts
@@ -5,7 +5,7 @@ export const DataSidebarMenu = [
     routePath: '/',
   },
   {
-    title: 'testes cadastrados',
+    title: 'testes realizados',
     iconName: 'chemical',
     routePath: '/tiposdeteste',
   },
@@ -13,15 +13,5 @@ export const DataSidebarMenu = [
     title: 'pessoas cadastradas',
     iconName: 'peoples',
     routePath: '/pessoas',
-  },
-  {
-    title: 'iniciar teste',
-    iconName: 'rocket',
-    routePath: '',
-  },
-  {
-    title: 'manual',
-    iconName: 'question1',
-    routePath: '/help',
   }
 ]
