--- src/App/View/Components/AverageGraph/_styles.scss
@@ -1,5 +1,0 @@
-canvas.average_graph {
-    padding-left: 80px;
-    padding-right: 80px;
-    padding-bottom: 90px;
-}
