--- src/App/View/Components/HelpList/_styles.scss
@@ -1,45 +1,0 @@
-.help_title{
-
-    padding-top: 30px;
-      
-    .wrapper{
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: center;
-    }
-  
-    h1, h2{ text-align: center }
-  
-    h3{
-      text-align: center;
-      margin: 5px;
-    }
-}
-  
-h2.help_info {
-    text-align: center;
-}
-  
-h3.help_info {
-    padding-left: 80px;
-    padding-right: 80px;
-}
-  
-.help_title_final{
-    padding-top: 30px;
-    padding-bottom: 90px;  
-    .wrapper{
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: center;
-    }
-  
-    h1, h2{ text-align: center }
-  
-    h3{
-      text-align: center;
-      margin: 5px;
-  
-      span{ font-weight: 400 }
-    }
-}
