+++ src/App/View/Components/ListItens/_styles.scss
@@ -1,8 +1,6 @@
 .list-itens{
   padding-top: 10px;
-  padding-bottom: 90px;
-  padding-left: 80px;
-  padding-right: 80px;
+  padding-bottom: 30px;
 
   ul {
     display: flex;
