+++ src/App/View/Components/SidebarMenu/_styles.scss
@@ -35,21 +35,11 @@ aside.sidebar_menu {
   .menu:hover{ transform: scale(1.2) }
 }
 
-// 80 - 5.5
-// 60 - 4,125
-// 25 - 0,38125
-
 aside.sidebar_menu.uncompressed{
-  position: fixed;
   width: 80px;
-  // left: 0;
-  // bottom: 50%;
-  left: 50%;
-  bottom: 0px;
-  transform: translateY(50%);
-  padding-top: 70px;
-  rotate: 270deg;
-  transform-origin: bottom left;
+  left: 0;
+  top: 50%;
+  transform: translateY(-50%);
 
   ul {
     border-radius: 0 30px 30px 0;
@@ -63,11 +53,8 @@ aside.sidebar_menu.uncompressed{
         flex-direction: column;
         justify-content: space-around;
         height: 60px;
-        rotate: 90deg;
   
-        .icon{ 
-          width: 25px;
-        }
+        .icon{ width: 25px }
         
         h3{
           color: $white;
