+++ src/App/View/Components/StartTestFixed/_styles.scss
@@ -1,0 +1,45 @@
+.start_test {
+  position: fixed;
+  bottom: 15px;
+  right: 15px;
+  z-index: 1000;
+
+  .wrapper_button{
+    border-radius: 50px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    transform: translateX(calc(100% - 35px));
+    
+    .icon {
+      background-color: $purple;
+      border-radius: 50px;
+      width: 35px;
+      padding: 10px;
+      height: 35px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+    }
+    
+    span {
+      opacity: 0;
+      color: $white;
+      padding-right:15px;
+      transition: 0.4s;
+    }
+
+    transition: transform 0.4s, background-color 0.3s;    
+  }
+
+  .wrapper_button:hover {
+    transform: unset;
+    background-color: $purple;
+
+    span {
+      position: unset;
+      opacity: 1;
+    }
+  }
+}
