+++ src/App/View/Routes.tsx
@@ -9,7 +9,6 @@ import TestGoingOn from "./Scenes/TestGoingOn";
 import Test from "./Scenes/Tests/Childs/Test";
 import Person from "./Scenes/Persons/Childs/Person";
 import UniqueTest from "./Scenes/UniqueTest";
-import Help from './Scenes/Help'
 
 export default class Routes extends React.Component {
   render() {
@@ -22,7 +21,6 @@ export default class Routes extends React.Component {
         <Route path="/pessoas/:peopleId" component={Person} />
         <Route path="/testando" component={TestGoingOn} />
         <Route path="/teste/:testId" component={UniqueTest} />
-        <Route path="/help" component={Help} />
       </Switch>
     )
   }
