+++ src/App/View/Scenes/Tests/Childs/Test/_style.scss
@@ -16,35 +16,3 @@
     span{ font-weight: 400 }
   }
 }
-
-h2.info {
-  text-align: center;
-}
-
-h3.info {
-  padding-left: 80px;
-  padding-right: 80px;
-}
-
-.info_final{
-  padding-top: 30px;
-  padding-bottom: 90px;  
-  .wrapper{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-  }
-
-  h1, h2{ text-align: center }
-
-  h3{
-    text-align: center;
-    margin: 5px;
-
-    span{ font-weight: 400 }
-  }
-}
-
-h2.testes_realizados_titulo {
-  text-align: center;
-}
