+++ src/App/View/Scenes/Tests/index.tsx
@@ -26,7 +26,6 @@ export default class Tests extends React.Component<Props, State> {
     let dataList: Array<itemOfListItensComponent> = [];
 
     
-    // eslint-disable-next-line array-callback-return
     testTypes.map((testType) => {
       dataList.push({
         title: testType.title,
@@ -41,10 +40,7 @@ export default class Tests extends React.Component<Props, State> {
 
   render() {
     return (
-      <div>
-        <h2 className="info">Testes Cadastrados</h2>
-        <ListItens data={this.state.dataList} />
-      </div>
+      <ListItens data={this.state.dataList} />
     )
   }
 }
