+++ src/App/View/index.tsx
@@ -11,6 +11,7 @@ import { DataSidebarMenu } from "../Data/SidebarMenu";
 //Components
 import SidebarMenu from "./Components/SidebarMenu";
 import SidebarNew from "./Components/SidebarNew";
+import StartTestFixed from "./Components/StartTestFixed";
 import Iframe from "./Components/Iframe";
 import FinishTestButton from "./Components/FinishTestButton";
 
@@ -50,7 +51,8 @@ class App extends React.Component<Props> {
 
         <div className={`${testGoingOn ? "invisible" : ""}`}>
           <SidebarNew data={DataSidebarNew} />
-          <SidebarMenu currentUrl={this.props.currentUrl} testGoingOn={this.props.testGoingOn} data={DataSidebarMenu} />
+          <SidebarMenu data={DataSidebarMenu} />
+          <StartTestFixed />
           <Routes />
         </div>
